scripts/requirements_rdi.txt
redis>=4.0.0
psycopg2-binary>=2.9.0
flask>=2.0.0
pandas>=1.3.0
sqlalchemy>=1.4.0
